refactor: extract multer config to middleware/upload.js
